Address PR comments - Single base class
