+++ packages/nimble-components/src/rich-text-mention/mention-base/models/mention-internals.ts
