+++ packages/nimble-components/src/rich-text-mention/mention-base/template.ts
@@ -1,7 +1,7 @@
 import { slotted, html } from '@microsoft/fast-element';
-import type { RichTextMentionBase, RichTextMentionConfig } from '.';
-import type { RichTextMentionBaseValidator } from './models/rich-text-mention-base-validator';
+import type { RichTextMention, RichTextMentionConfig } from '.';
+import type { RichTextMentionValidator } from './models/rich-text-mention-base-validator';
 
 export const template = html<
-RichTextMentionBase<RichTextMentionConfig, RichTextMentionBaseValidator<[]>>
+RichTextMention<RichTextMentionConfig, RichTextMentionValidator<[]>>
 >`<slot ${slotted('mappings')} name="mapping"></slot>`;
